+++ test/sep0012_test.dart
@@ -2,7 +2,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:stellar_flutter_sdk/stellar_flutter_sdk.dart';
 import 'package:http/http.dart' as http;
-import 'package:http_parser/http_parser.dart';
 import 'package:http/testing.dart';
 import 'dart:convert';