+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.7.5] - 23.Feb.2024.
+- update SEP-12 financial account fields
+- update SEP-06 to reflect current SEP doc version
+
 ## [1.7.4] - 01.Feb.2024.
 - add SEP-38 support